switch to int8_t when applicable

+++ game/player.h
@@ -12,7 +12,7 @@
 typedef struct
 {
     char id[32];                // Unique ID for the entity type
-    int index;                  // Index for the specific entity instance
+    int8_t index;               // Index for the specific entity instance
     Vector size;                // Size of the entity
     Sprite *sprite_right;       // Entity sprite when looking right
     Sprite *sprite_left;        // Entity sprite when looking left
@@ -26,7 +26,7 @@ typedef struct
     float speed;                // Speed of the entity
     float attack_timer;         // Cooldown duration between attacks
     float elapsed_attack_timer; // Time elapsed since the last attack
-    float strength;             // Damage the entity deals
+    int8_t strength;            // Damage the entity deals
     float health;               // Health of the entity
     char message[64];           // Message to display when interacting with the entity
     bool is_user;               // Flag to indicate if the entity is a live player or not
@@ -79,10 +79,10 @@ typedef struct
     Entity *player;
     //
     float fps;
-    int level_count;
-    int enemy_count;
-    int npc_count;
-    int current_level;
+    int8_t level_count;
+    int8_t enemy_count;
+    int8_t npc_count;
+    int8_t current_level;
     bool ended_early;
     Imu *imu;
     bool imu_present;
@@ -98,8 +98,8 @@ typedef struct
     //
     GameMode game_mode;
     //
-    int icon_count;
-    int icon_offset;
+    uint32_t icon_count;
+    uint16_t icon_offset;
     //
     char message[64];
     //
